Revert "Migrate from MightyPirates to MovingBlocks JNLua [phase 1]"
This reverts commit b426a7f5638e4f3278b005a4de866b6050520e49.

+++ src/main/scala/li/cil/oc/server/machine/luac/UnicodeAPI.scala
@@ -45,7 +45,7 @@ class UnicodeAPI(owner: NativeLuaArchitecture) extends NativeLuaAPI(owner) {
         })
         else string.length
       if (end <= start) lua.pushString("")
-      else lua.pushString(string.substring(start.toInt, end.toInt))
+      else lua.pushString(string.substring(start, end))
       1
     })
     lua.setField(-2, "sub")
